requirements.txt
dotenv==0.9.9
requests==2.32.5
pyspark==3.5.0
psycopg2-binary==2.9.11
apache-airflow-providers-apache-spark